9/25: Created this class to text to see if the database class just has a problem with calling methods and as I assumed that was the case
this class works with calling methods from another activity now what next is to make an api class where all the api calls are gonna be made in. Then Try to make the database class work or make a new database class only of activity calls to run the methods in another class. Also use multithreading so this application does not bog down the CPU and cause the application
 to freeze. If SQLlite database does not work then look into firebase